+++ packages/adapters/rig/helius/src/create_webhook.rs
@@ -12,17 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    actions::{create_webhook, HeliusWebhookResponse},
-    parameters_json_schema, SolanaAgentKit,
-};
-use rig::{
-    completion::ToolDefinition,
-    tool::{Tool, ToolEmbedding},
-};
-use serde::{Deserialize, Serialize};
-use serde_json::json;
+use solagent_core::{rig::{completion::ToolDefinition, tool::{Tool, ToolEmbedding}}, SolanaAgentKit, parameters_json_schema};
 use std::sync::Arc;
+use serde::{Deserialize, Serialize};
+use solagent_plugin_helius::{create_webhook, HeliusWebhookResponse};
 
 #[derive(Deserialize)]
 pub struct CreateWebHookArgs {
