+++ packages/adapters/rig/helius/src/lib.rs
@@ -17,3 +17,4 @@ pub mod delete_webhook;
 pub mod get_assets_by_owner;
 pub mod get_webhook;
 pub mod transaction_parsing;
+
