+++ packages/plugins/helius/src/get_webhook.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::SolanaAgentKit;
+use solagent_core::SolanaAgentKit;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
