+++ packages/plugins/helius/src/transaction_parsing.rs
@@ -12,13 +12,30 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::SolanaAgentKit;
+use solagent_core::SolanaAgentKit;
+use serde::{Deserialize, Serialize};
 use serde_json::json;
 
-pub async fn get_assets_by_owner(
+#[derive(Debug, Serialize, Deserialize)]
+pub struct HeliusWebhookIdResponse {
+    pub wallet: String,
+    pub webhook_url: String,
+    pub transaction_types: Vec<String>,
+    pub account_addresses: Vec<String>,
+    pub webhook_type: String,
+}
+
+/// Parse a Solana transaction using the Helius Enhanced Transactions API
+///
+/// # Arguments
+/// * `agent` - An instance of SolanaAgentKit (with .config.HELIUS_API_KEY)
+/// * `transaction_id` - The transaction ID to parse
+///
+/// # Returns
+/// Parsed transaction data
+pub async fn transaction_parse(
     agent: &SolanaAgentKit,
-    owner_public_key: &str,
-    limit: u32,
+    transaction_id: &str,
 ) -> Result<serde_json::Value, Box<dyn std::error::Error>> {
     // Get the Helius API key from the agent's configuration
     let api_key = match agent.config.helius_api_key.as_ref() {
@@ -26,34 +43,15 @@ pub async fn get_assets_by_owner(
         None => return Err("Missing Helius API key in agent.config.HELIUS_API_KEY".into()),
     };
 
-    let url = format!("https://mainnet.helius-rpc.com/?api-key={}", api_key);
-
     let client = reqwest::Client::new();
+    let url = format!("https://api.helius.xyz/v0/transactions/?api-key={}", api_key);
 
-    let request_body = json!({
-        "jsonrpc": "2.0",
-        "id": "get-assets",
-        "method": "getAssetsByOwner",
-        "params": json!({
-            "ownerAddress": owner_public_key,
-            "page": 3,
-            "limit": limit,
-            "displayOptions": { "showFungible": true },
-        }),
+    let body = json!( {
+        "transactions": vec![transaction_id.to_string()],
     });
 
-    let response = client.post(&url).header("Content-Type", "application/json").json(&request_body).send().await?;
+    let response = client.post(url).header("Content-Type", "application/json").json(&body).send().await?;
 
-    if !response.status().is_success() {
-        return Err(format!(
-            "Failed to fetch: {} - {}",
-            response.status(),
-            response.status().canonical_reason().unwrap_or("Unknown")
-        )
-        .into());
-    }
-
-    let data: serde_json::Value = response.json().await?;
-
+    let data = response.json().await?;
     Ok(data)
 }
